fix: moved registeration components to components folder
